requirements.txt
click
colour
pillow

